// models/input/paging-options-params.interface.ts
/**
 * Вспомогательный интерфейс описывающий опции при запросе постраничного вывода данных
 */
export interface IPagingOptionsParams {
	/**
	 * 	Номер страницы (zero-based)
	 */
	pageNumber: number;
	/**
	 * Размер страницы, должен быть 0 < pageSize <= 100	Если размер страницы не указан, то система считает pageSize = 20 (значение по умолчанию)
	 */
	pageSize: number;
}